requirements.txt
# Core dependencies for FinCLI

# CLI Framework
typer[all]==0.12.3
rich==13.7.1

# Google APIs
google-api-python-client==2.134.0
google-auth-oauthlib==1.2.0
google-auth-httplib2==0.2.0

# LLM Providers
boto3==1.34.144  # AWS Bedrock
requests==2.32.3  # Ollama HTTP client
openai>=1.0.0  # OpenAI GPT models (optional)
anthropic>=0.18.0  # Anthropic Claude API (optional)

# Database
SQLAlchemy==2.0.31
alembic==1.13.2

# Configuration
pydantic==2.8.2
pydantic-settings==2.3.4
python-dotenv==1.0.1

# Data Processing
pandas==2.2.2
python-dateutil==2.9.0.post0

# Logging
structlog==24.4.0

# Retry Logic
tenacity==8.5.0

# API Framework
fastapi==0.115.0
uvicorn[standard]==0.30.6
python-multipart==0.0.9
httpx==0.27.0  # For TestClient
